face-service/requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
insightface==0.7.3
onnxruntime==1.22.1
numpy==1.26.4
pillow==10.4.0
faiss-cpu==1.12.0
python-multipart==0.0.9
pydantic==2.8.2
